% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_hunt.R
\name{color_hunt}
\alias{color_hunt}
\title{color_hunt}
\usage{
color_hunt(url)
}
\arguments{
\item{url}{A url that you want to import the colors.}
}
\description{
This function make palette from https://dribbble.com/
}
\examples{
library(ggplot2)
library(dplyr)
colors <- color_hunt("https://dribbble.com/shots/6393225-Iris")
iris \%>\%
ggplot(aes(Sepal.Length, Sepal.Width, color = Species)) +
geom_point(size = 2) +
scale_color_manual(values = colors[-1]) +
theme(
  panel.background = element_rect(fill = colors[1])
)
}
\keyword{color,}
\keyword{dribbble}
\keyword{ggplot2,}
