% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_lego.R
\name{color_lego}
\alias{color_lego}
\title{color_lego}
\usage{
color_lego(series)
}
\arguments{
\item{series}{A series that you want to import the colors. The list of series can find at colorhount::list_lego}
}
\description{
This function make palette from lego color dataset
}
\examples{
colors <- color_lego("Ferrari")
mpg \%>\%
group_by(class) \%>\%
  summarise(displ = mean(displ)) \%>\%
  ggplot(aes(reorder(class, displ), displ, fill = class)) +
  geom_col() +
  scale_fill_manual(values = color_lego("Ferrari"))
}
\keyword{color,}
\keyword{ggplot2,}
\keyword{lego}
