% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_lego_show.R
\name{color_lego_show}
\alias{color_lego_show}
\title{color_lego_show}
\usage{
color_lego_show(series)
}
\arguments{
\item{series}{A series that you want to import the colors. The list of series can find at colorhount::list_lego}
}
\description{
This function shows the colors of palette from lego color dataset
}
\examples{
color_lego_show("Ferrari")
color_lego_show(sample(list_lego, 1))
}
\keyword{color,}
\keyword{ggplot2,}
\keyword{lego}
